pyproject.toml
[tool.poetry]
name = "fastapi-boilerplate"
version = "0.1.0"
description = ""
authors = ["Dyarlen Iber <dyarlen1@gmail.com>"]
readme = "README.md"
packages = [{include = "fastapi_boilerplate"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.0"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
python-dotenv = "^1.0.0"
httpx = "^0.24.1"
anyio = "^3.7.1"
pymongo = "^4.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.3.0"
taskipy = "^1.11.0"
isort = "^5.12.0"

[tool.pytest.ini_options]
pythonpath = "."
minversion = "6.0"
testpaths = [
    "tests"
]

[tool.isort]
profile = "black"

[tool.taskipy.tasks]
dev = "uvicorn fastapi_boilerplate.main:app --reload"
lint = "black . && isort ."
lint_check = "black --check --diff . && isort --check --diff ."
pre_test = "task lint_check"
test = "pytest -s -x --cov=fastapi_boilerplate -vv"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
